odsi_app/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bleach==6.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
Flask==3.1.2
Flask-Limiter==4.1.1
Flask-Session==0.8.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
MarkupSafe==3.0.3
msgspec==0.20.0
numpy==2.3.3
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pycparser==2.23
pycryptodome==3.23.0
pycryptodomex==3.23.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.1.0
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.4
wrapt==2.0.1
